Organized code in a better way. I may add more to the app for fun this weekend, but it should match what I used in my video.
