+++ TutorNotes/HomeScreen.xaml.cs
@@ -196,7 +196,6 @@ namespace TutorNotes
                 return;
             }
 
-
             if (((Educator)App.CurrentUser).StudentsAssigned.Count != 0)
             {
                 AddSessionWindow addSessionWindow = new AddSessionWindow(cellInfo, this);
@@ -221,62 +220,6 @@ namespace TutorNotes
                 cell.CellName = "";
                 cell.Emptiness = true;
             }
-        }
-    }
-    public class CellInfo : INotifyPropertyChanged
-    {
-        private Brush _fill;
-        public bool _emptiness;
-        public string CellName { get; set; }
-        public Brush BorderBrush { get; set; }
-        public Thickness BorderThickness { get; set; }
-
-        public CellInfo(string name, Brush borderBrush, Thickness borderThickness, int row, int col, Brush fill)
-        {
-            CellName = name;
-            BorderBrush = borderBrush;
-            BorderThickness = borderThickness;
-            Fill = fill;
-            Emptiness = true; // Will initially be empty when setting up the program.
-        }
-
-        public bool Emptiness
-        { 
-            get
-            {
-                return this._emptiness;
-            }
-            set
-            {
-                if (_emptiness != value)
-                {
-                    _emptiness = value;
-                    OnPropertyChanged(nameof(Emptiness));
-                }
-            }
-        }
-
-        public Brush Fill
-        {
-            get
-            {
-                return this._fill;
-            }
-            set
-            {
-                if (this._fill != value)
-                {
-                    this._fill = value;
-                    OnPropertyChanged(nameof(Fill));
-                }
-            }
-        }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        protected void OnPropertyChanged(string propertyName)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
     }
 }
