+++ TutorNotes/MainWindow.xaml
@@ -95,7 +95,6 @@
 
             <!-- Image Design -->
             <Image Source="/Images/TutorNotesMain.png" Grid.Row="1" Grid.Column="0" Grid.ColumnSpan="4" Grid.RowSpan="7" Margin="0,0,50,0"  Stretch="Fill"/>
-
             <Rectangle x:Name="RectangleHorizontal" Fill="Transparent" Grid.Column="3" Grid.Row="1" Grid.ColumnSpan="6" MinHeight="195" MaxHeight="1000" MinWidth="200" MaxWidth="400"/>
             <Rectangle x:Name="RectangleVertical" Fill="Transparent" Grid.Column="3" Grid.Row="1" Grid.RowSpan="7" MinWidth="240" MaxWidth="300" MinHeight="400" MaxHeight="1000"/>
 
@@ -138,13 +137,10 @@
                 </Button>
             </Viewbox>
 
-            
             <!-- Create Account/ Forgot Password -->
-
-                <TextBlock x:Name="createAccountTxt" Grid.Column="4" Grid.Row="5" HorizontalAlignment="Left" VerticalAlignment="Center" FontSize="10" Margin="5" Text="Create account"  Foreground="{StaticResource CustomColor1}" MouseEnter="createAccount_MouseEnter" MouseLeave="createAccountTxt_MouseLeave" MouseDown="createAccountTxt_MouseDown" MouseLeftButtonDown="createAccountTxt_MouseLeftButtonDown">
-                </TextBlock>
-
-                <TextBlock x:Name="forgotPasswordTxt" Grid.Column="5" Grid.Row="5" HorizontalAlignment="Right"  VerticalAlignment="Center" FontSize="10" Margin="5" Text="Forgot password?" Foreground="{StaticResource CustomColor1}" MouseEnter="forgotPasswordTxt_MouseEnter" MouseLeave="forgotPasswordTxt_MouseLeave" MouseLeftButtonDown="forgotPasswordTxt_MouseLeftButtonDown"/>
+            <TextBlock x:Name="createAccountTxt" Grid.Column="4" Grid.Row="5" HorizontalAlignment="Left" VerticalAlignment="Center" FontSize="10" Margin="5" Text="Create account"  Foreground="{StaticResource CustomColor1}" MouseEnter="createAccount_MouseEnter" MouseLeave="createAccountTxt_MouseLeave" MouseDown="createAccountTxt_MouseDown" MouseLeftButtonDown="createAccountTxt_MouseLeftButtonDown">
+            </TextBlock>
+            <TextBlock x:Name="forgotPasswordTxt" Grid.Column="5" Grid.Row="5" HorizontalAlignment="Right"  VerticalAlignment="Center" FontSize="10" Margin="5" Text="Forgot password?" Foreground="{StaticResource CustomColor1}" MouseEnter="forgotPasswordTxt_MouseEnter" MouseLeave="forgotPasswordTxt_MouseLeave" MouseLeftButtonDown="forgotPasswordTxt_MouseLeftButtonDown"/>
 
         </Grid>
     </Border>
