+++ TutorNotes/UpdateInfoWindow.xaml
@@ -5,7 +5,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:TutorNotes" Icon="Images/writing.ico"
         mc:Ignorable="d" FontSize="18" FontFamily="Century Gothic"  WindowStartupLocation="CenterScreen"
-        Title="Tutor Notes" Height="350" Width="450" WindowStyle="None" ResizeMode="CanResizeWithGrip" 
+        Title="Tutor Notes" Height="380" Width="450" WindowStyle="None" ResizeMode="CanResizeWithGrip" 
         DataContext="{Binding RelativeSource={RelativeSource Self}}" Background="{x:Null}" MouseLeftButtonDown="dragWindow" >
 
     <WindowChrome.WindowChrome>
@@ -98,13 +98,13 @@
                 <ComboBoxItem Content="D" />
                 <ComboBoxItem Content="F" />
             </ComboBox>
-            <TextBlock Grid.Column="2" Grid.Row="4" Grid.ColumnSpan="5" FontFamily="Century Gothic" Text="Current Topic" Margin="0" HorizontalAlignment="Center" VerticalAlignment="Bottom" FontSize="25" Foreground="{StaticResource CustomColor4}"/>
-            <TextBox x:Name="updateCurrentTopic" Grid.Row="5" Grid.Column="2" Grid.ColumnSpan="5" Width="250" Height="30" Margin="0" HorizontalAlignment="Center" VerticalAlignment="Top"/>
+            <TextBlock Grid.Column="2" Grid.Row="4" Grid.ColumnSpan="5" FontFamily="Century Gothic" Text="Current Topic" Margin="0, 10, 0, 0" HorizontalAlignment="Center" VerticalAlignment="Top" FontSize="25" Foreground="{StaticResource CustomColor4}"/>
+            <TextBox x:Name="updateCurrentTopic" Grid.Row="4" Grid.RowSpan="2" Grid.Column="2" Grid.ColumnSpan="5" Width="250" Height="30" Margin="0, 40, 0, 0" HorizontalAlignment="Center" VerticalAlignment="Top"/>
 
-            <TextBlock Grid.Column="2" Grid.Row="6" Grid.ColumnSpan="5" FontFamily="Century Gothic" Text="Next Possible Topics" Margin="0" HorizontalAlignment="Center" VerticalAlignment="Top" FontSize="25" Foreground="{StaticResource CustomColor4}"/>
-            <TextBox x:Name="updateNextTopics" Grid.Row="6" Grid.Column="2" Grid.RowSpan="2" Grid.ColumnSpan="5" Width="250" Height="30" Margin="0, 30, 0, 0" HorizontalAlignment="Center" VerticalAlignment="Top"/>
+            <TextBlock Grid.Column="2" Grid.Row="5" Grid.ColumnSpan="5" FontFamily="Century Gothic" Text="Next Possible Topics" Margin="0, 20, 0, 0" HorizontalAlignment="Center" VerticalAlignment="Top" FontSize="25" Foreground="{StaticResource CustomColor4}"/>
+            <ListBox x:Name="updateNextTopics" Grid.Row="5" Grid.Column="2" Grid.RowSpan="3" Grid.ColumnSpan="5" Width="250" Height="75" Margin="0, 0, 0, 0" HorizontalAlignment="Center" VerticalAlignment="Center"/>
 
-            <Button x:Name="saveStudentEditsBttn" Content="Save" Grid.Column="6" Grid.Row="7" FontSize="15" Height="25" Width="50" Margin="5, 10, 0, 0" HorizontalAlignment="Left" VerticalAlignment="Top"  Foreground="{StaticResource CustomColor4}" Background="{StaticResource CustomColor6}" BorderBrush="Transparent" FontFamily="Century Gothic" FontWeight="Bold" Click="saveStudentEditsBttn_Click">
+            <Button x:Name="saveStudentEditsBttn" Content="Save" Grid.Column="6" Grid.Row="7" FontSize="15" Height="25" Width="50" Margin="5, 15, 0, 0" HorizontalAlignment="Left" VerticalAlignment="Top"  Foreground="{StaticResource CustomColor4}" Background="{StaticResource CustomColor6}" BorderBrush="Transparent" FontFamily="Century Gothic" FontWeight="Bold" Click="saveStudentEditsBttn_Click">
                 <Button.Resources>
                     <Style TargetType="Border" BasedOn="{StaticResource RoundMe}"/>
                 </Button.Resources>
